pyproject.toml
[project]
name = "ptt-llm"
version = "0.1.0"
description = "PTT Stock Daily Pulse Analyst"
authors = [
    {name = "wise4466",email = "91211wise@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "langchain-google-genai (>=4.0.0,<5.0.0)",
    "langchain-pinecone (>=0.2.13,<0.3.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "pinecone (<8.0.0)",
    "langchain (>=1.1.3,<2.0.0)",
    "langchain-core (>=1.2.0,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
